(FRONT_CreateDay) Move day creation to separate button
