+++ PSK.UII/ClientApp/src/components/Home/HomePage.js
@@ -1,12 +1,7 @@
 import React from 'react';
-import { connect } from 'react-redux';
 import Calendar from '../Calendar/Calendar';
 
 export default class HomePage extends React.Component {
-    constructor(props) {
-        super(props);
-    }
-
     render() {
         return (
             <React.Fragment>
